backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
beautifulsoup4
requests
python-dotenv

google-genai
langchain-core
langchain-community
langchain-google-genai
pydantic
pydantic-settings